==> Part 1 - Clinical Data Preprocessing and Sample Selection (1) (1).ipynb <==
# Data Preprocessing:

In this step the entire clinical dataset containing all information regarding TCGA-LUAD patients is narrowed down. The idea is to create 2 groups of patients based on their smoking status. Group 1 consists of 10 randomly selected patients (Male or Female) for which years_smoked data is available. 

The same steps are followed for the second group. The second group consists of 10 randomly selected patients (Male or Female) for whom smoking exposure is not available, and thus are considered different from smoking group/group 1.

After identifying the patients of interest the submitter_id of the patients is then used to obtain those patients' DNA Methylation data from GDC.


import pandas as pd #Necessary package for importing the datasets.

clinicaldata = pd.read_csv("TCGA-LUAD_Clinical.csv") #Entire clinical dataset 
clinicaldata.head() #Checking the dataset looks right.

clinicaldata.shape #Making sure its complete.

for col in clinicaldata.columns: #The names of all columns in clinical dataset.
    print(col)


There are a lot of columns here that we wont be using right now. For simplicity purposes, lets reduce the columns down to the ones we need right now. 

clindata_copy = clinicaldata[['submitter_id', 'cigarettes_per_day', 'alcohol_intensity', 
                              'years_smoked','alcohol_history', 'pack_years_smoked', 
                              'gender', 'race', 'vital_status', 'ethnicity', 
                              'bcr_patient_barcode','disease']].copy()
#Leaving in some columns to get a good overview of the data.

clindata_copy.shape

The dataset is now 13 columns instead of 74. Much easier to understand and process.

# Group 1: Smokers

Below the data is being sorted by most years smoked to lowest years smoked. This is the dataset from which group 1 patients are randomly selected.

smokersdata = clindata_copy.nlargest(200, 'years_smoked') 
#Saving the data that contains all 200 samples.
#The dataset below ranges from 64 years smoked to 2 years smoked.

smokersdata.head()

Group1 = smokersdata.sample(n = 10) #Randomly selecting the 10 patients from group 1.

Group1

Due to the nature of sample, I must save group 1 data to a csv. Otherwise, when line 9 is run Group 1 will become a completely new, random, set of 10 patients.

Group1.to_csv('group1.csv') #Saving Group 1 data into a csv.

Now group 1 is a file containing all the Ids of 10 smoker patients who have lung cancer. It is a mix of male and female patients. The patient's submitter_id will be used to obtain their DNA Methylation file.

# Group 2: Non-Smokers

In order to appropriately select 10 random patients for this group, for each of the 10 patients, all exposure data must be zero. The data currently contains values called NaN. These values must be replaced with 0. After the data is clean we select the 10 random patients. 

clindata_copy = clindata_copy.fillna(0) #Replaces all NaNs with 0.

clindata_copy.describe() #Take a look at the data. There are no NaNs visible.

clindata_copy.isna().sum() #This counts the # of NaNs in each column. It verifies there are no NaN in the dataset.

Now that the dataset is clean we select all patients for which cigarettes_per_day, years_smoked and pack_years_smoked is equal to zero.

nonsmokers = clindata_copy.loc[(clindata_copy['cigarettes_per_day'] == 0) & (clindata_copy['years_smoked'] == 0) & (clindata_copy['pack_years_smoked'] == 0)]
#This selects all data for which the exposure data is equal to zero.

nonsmokers.describe() #We verify that for those columns the max values are 0.

Group2 = nonsmokers.sample(n = 10) #Randomly selecting the 10 patients from group 2.

Group2 #Seeing what the data looks like.

Group2.to_csv('group2.csv') #Saving Group 2 as csv.

For organizational purposes this notebook ends here. Notebook 2 consists of analyzing DNA Methylation data of the patients we selected here.
