==> Part 2 - Group 1 Methylation Data Preprocessing and Download (1).ipynb <==
# I. Introduction

In this notebook DNA Methylation data is obtained manually from GDC by searching for the selected patient using their id (submitter_id). After obtaining the data for each patient, the data is then cleaned and downloaded to further explore in notebook 4.


# DNA Methylation Data Gathering Steps

This is mostly for records keeping. I am adding it in case you wanted to see how exactly I obtained the DNA Methylation data for every patient.

Steps followed to obtain DNA Methylation Data:

1. Go to gdc.cancer.gov (Its where all TCGA data is stored.)
2. Click on Lauch Data Portal.
3. Click Accept if asked.
4. Click on the green Projects button.
5. Under 'Primary Site' select 'Bronchus and Lung' and under 'Program' select 'TCGA'.
6. Click on the project name TCGA-LUAD.
7. Click on Explore Project Data button.
8. In the search bar on the left, enter the submitter_id of the patient. Do not hit enter.
9. From the dropdown results, select the patient.
10. Click on the case id.
11. Scroll to 'File Counts by Data Category', Find 'DNA Methylation' and click on the file count (usually a 1).
12. Click on the file name (.txt)
13. Click on Download.
14. Unzip, extract and save to intended folder.


# Group 1: Smoking Group

import pandas as pd #Necessary package/s.

group1 = pd.read_csv("group1.csv") #import data

group1

For every submitter_id in the group 1 dataset above, I follow the data gathering steps. They will be labeled DMG1P1 :DNA Methylation Group 1 patient 1. Patient 1 is the first one on the list therefore, DMG1P1 = TCGA-MP-A4TE's DNA Methylation data. For group 2 the same naming convention is followed, only change is group # (e.g. DMG2P1).

#Loading in all 10 patients from group 1 (smokers group)
g1p1 = pd.read_csv('DMG1P1.txt',  delimiter = '\t', encoding = 'utf-8')
g1p2 = pd.read_csv('DMG1P2.txt',  delimiter = '\t', encoding = 'utf-8')
g1p3 = pd.read_csv('DMG1P3.txt',  delimiter = '\t', encoding = 'utf-8')
g1p4 = pd.read_csv('DMG1P4.txt',  delimiter = '\t', encoding = 'utf-8')
g1p5 = pd.read_csv('DMG1P5.txt',  delimiter = '\t', encoding = 'utf-8')
g1p6 = pd.read_csv('DMG1P6.txt',  delimiter = '\t', encoding = 'utf-8')
g1p7 = pd.read_csv('DMG1P7.txt',  delimiter = '\t', encoding = 'utf-8')
g1p8 = pd.read_csv('DMG1P8.txt',  delimiter = '\t', encoding = 'utf-8')
g1p9 = pd.read_csv('DMG1P9.txt',  delimiter = '\t', encoding = 'utf-8')
g1p10 = pd.read_csv('DMG1P10.txt',  delimiter = '\t', encoding = 'utf-8')

# DNA Methylation Data Cleaning
## Group 1 - Patient 1

g1p1.head()#This is what the DNA Methylation Data for this patient looks like.

g1p1.isna().sum() #Lets see how much of the data is missing and for which column.

There seems to be quite a few missing values in the Beta_value column. Lets find out what percentage of the data is missing.

len(g1p1['Beta_value']) #Total count for Beta_value. All datasets have the same total.

print((90207 / 485577) * 100 , '%') #Percent of missing Beta_value data.

There may be a better way of handling missing values for this dataset but for now I will be removing the rows for which Beta_value is not available. Another way to handle these values would be to set them to 0, but that would not be a good idea with this type of data because it would give those rows a meaning they truly dont have. This is due to the Beta_value ranging from 0 to 1.

g1p1 = g1p1.dropna() #Removing rows of missing values.

g1p1.isna().sum() #Checking that the data is clean and free from NAs.

len(g1p1['Beta_value']) #This is to make sure the values removed are the missing ones and nothing else.

g1p1.head()

Lets do the same data cleaning as above to every patient on group 1. 

## Group 1 - Patient 2

g1p2.head() #Lets see the data.

g1p2.isna().sum() #Total sum of all missing values for every column.

g1p2 = g1p2.dropna() #Removing rows of missing values.

g1p2.isna().sum() #Checking that the data is clean and free from NAs.

There seems to be a trend amongst the mising values. Perhaps all datasets have approximately the same value missing from the same column? If so, we can skip a few steps and go straight to removal of missing values.

## Group 1 - Patient 3

g1p3.isna().sum() #Total sum of all missing values for every column.

g1p3 = g1p3.dropna() #Removing rows of missing values.

g1p3.isna().sum() #Checking that the data is clean and free from NAs.

## Group 1 - Patient 4

g1p4.isna().sum() #Total sum of all missing values for every column.

g1p4 = g1p4.dropna() #Removing rows of missing values.

g1p4.isna().sum() #Checking that the data is clean and free from NAs.

## Group 1 - Patient 5

g1p5.isna().sum() #Total sum of all missing values for every column.

g1p5 = g1p5.dropna() #Removing rows of missing values.

g1p5.isna().sum() #Checking that the data is clean and free from NAs.

## Group 1 - Patient 6

g1p6.isna().sum() #Total sum of all missing values for every column.

g1p6 = g1p6.dropna() #Removing rows of missing values.

g1p6.isna().sum() #Checking that the data is clean and free from NAs.

## Group 1 - Patient 7

g1p7.isna().sum() #Total sum of all missing values for every column.

g1p7 = g1p7.dropna() #Removing rows of missing values.

g1p7.isna().sum() #Checking that the data is clean and free from NAs.

## Group 1 - Patient 8

g1p8.isna().sum() #Total sum of all missing values for every column.

g1p8 = g1p8.dropna() #Removing rows of missing values.

g1p8.isna().sum() #Checking that the data is clean and free from NAs.

## Group 1 - Patient 9

g1p9.isna().sum() #Total sum of all missing values for every column.

g1p9 = g1p9.dropna() #Removing rows of missing values.

g1p9.isna().sum() #Checking that the data is clean and free from NAs.

## Group 1 - Patient 10

g1p10.isna().sum() #Total sum of all missing values for every column.

g1p10 = g1p10.dropna() #Removing rows of missing values.

g1p10.isna().sum() #Checking that the data is clean and free from NAs.

# Download All Clean Files:

Lets download all the patient's clean datasets.

g1p1.to_csv('g1p1.csv')
g1p2.to_csv('g1p2.csv')
g1p3.to_csv('g1p3.csv')
g1p4.to_csv('g1p4.csv')
g1p5.to_csv('g1p5.csv')
g1p6.to_csv('g1p6.csv')
g1p7.to_csv('g1p7.csv')
g1p8.to_csv('g1p8.csv')
g1p9.to_csv('g1p9.csv')
g1p10.to_csv('g1p10.csv')

Now that all DNA Methylation data for every patient on group 1 is clean, I will use it to perform exploratory data analysis in part 4.
