==> Part 3 - Group 2 Methylation Data Preprocessing and Download (1).ipynb <==
# I. Introduction

In this notebook DNA Methylation data is obtained manually from GDC by searching for the selected patient using their id (submitter_id). After obtaining the data for each patient, the data is then cleaned and downloaded to further explore in notebook 4.


# DNA Methylation Data Gathering Steps

This is mostly for records keeping. I am adding it in case you wanted to see how exactly I obtained the DNA Methylation data for every patient.

Steps followed to obtain DNA Methylation Data:

1. Go to gdc.cancer.gov (Its where all TCGA data is stored.)
2. Click on Lauch Data Portal.
3. Click Accept if asked.
4. Click on the green Projects button.
5. Under 'Primary Site' select 'Bronchus and Lung' and under 'Program' select 'TCGA'.
6. Click on the project name TCGA-LUAD.
7. Click on Explore Project Data button.
8. In the search bar on the left, enter the submitter_id of the patient. Do not hit enter.
9. From the dropdown results, select the patient.
10. Click on the case id.
11. Scroll to 'File Counts by Data Category', Find 'DNA Methylation' and click on the file count (usually a 1).
12. Click on the file name (.txt)
13. Click on Download.
14. Unzip, extract and save to intended folder.


# Group 2: Non - Smoking Group

import pandas as pd #Necessary package/s.

group2 = pd.read_csv("group2.csv") #import data

group2

For every submitter_id in the group 2 dataset above, I follow the data gathering steps. They will be labeled DMG2P1 :DNA Methylation Group 2 patient 1. Patient 1 is the first one on the list therefore, DMG2P1 = TCGA-55-8508's DNA Methylation data. 

#Loading in all 10 patients from group 2 (non-smokers group)
g2p1 = pd.read_csv('DMG2P1.txt',  delimiter = '\t', encoding = 'utf-8')
g2p2 = pd.read_csv('DMG2P2.txt',  delimiter = '\t', encoding = 'utf-8')
g2p3 = pd.read_csv('DMG2P3.txt',  delimiter = '\t', encoding = 'utf-8')
g2p4 = pd.read_csv('DMG2P4.txt',  delimiter = '\t', encoding = 'utf-8')
g2p5 = pd.read_csv('DMG2P5.txt',  delimiter = '\t', encoding = 'utf-8')
g2p6 = pd.read_csv('DMG2P6.txt',  delimiter = '\t', encoding = 'utf-8')
g2p7 = pd.read_csv('DMG2P7.txt',  delimiter = '\t', encoding = 'utf-8')
g2p8 = pd.read_csv('DMG2P8.txt',  delimiter = '\t', encoding = 'utf-8')
g2p9 = pd.read_csv('DMG2P9.txt',  delimiter = '\t', encoding = 'utf-8')
g2p10 = pd.read_csv('DMG2P10.txt',  delimiter = '\t', encoding = 'utf-8')

# DNA Methylation Data Cleaning
## Group 2 - Patient 1

g2p1.head()#This is what the DNA Methylation Data for this patient looks like.

g2p1.isna().sum() #Lets see how much of the data is missing and for which column.

There seems to be quite a few missing values in the Beta_value column. Lets find out what percentage of the data is missing.

#Total count for Beta_value. Should be the same for all datasets.
len(g2p1['Beta_value']) 

print((89776 / 485577) * 100 , '%') #Percent of missing Beta_value data.

There may be a better way of handling missing values for this dataset but for now I will be removing the rows for which Beta_value is not available. Another way to handle these values would be to set them to 0, but that would not be a good idea with this type of data because it would give those rows a meaning they truly dont have. This is due to the Beta_value ranging from 0 to 1.

g2p1 = g2p1.dropna() #Removing rows of missing values.

g2p1.isna().sum() #Checking that the data is clean and free from NAs.

#This is to make sure the values removed are the missing ones and nothing else.

len(g2p1['Beta_value']) 


485577 - 89776 #total - missing values = properly cleaned dataset total

Lets do the same data cleaning as above to every patient on group 2. 

## Group 2 - Patient 2

g2p2.head() #Lets see the data.

g2p2.isna().sum() #Total sum of all missing values for every column.

len(g2p2["Beta_value"]) #Total count.

The DNA Methylation dataset for this patient is much smaller (27578 vs 485577) than others from the same group. Thus I will select another patient for whom DNA methylation file is closer to the other patients. This is so that any comparisons made are fair.


## Choosing a New Patient 2 for Group 2

#All clinical data is imported to choose new pt for group 2.

clinicaldata = pd.read_csv("TCGA-LUAD_Clinical.csv") 

#There are too many columns. 
#Lets remove some and leave in some to get a good overview of the data.

clindata_copy = clinicaldata[['submitter_id', 'cigarettes_per_day', 
                              'alcohol_intensity', 'years_smoked',
                              'alcohol_history', 'pack_years_smoked', 
                              'gender', 'race', 'vital_status', 'ethnicity', 
                              'bcr_patient_barcode','disease']].copy()

clindata_copy = clindata_copy.fillna(0) #Replaces all NaNs with 0.

#This selects all data for which the exposure data is equal to zero.
nonsmokers = clindata_copy.loc[(clindata_copy['cigarettes_per_day'] == 0) 
                               & (clindata_copy['years_smoked'] == 0) 
                               & (clindata_copy['pack_years_smoked'] == 0)]

#Randomly selects 1 patient that meets the above requirements.
NewPatient2 = nonsmokers.sample(n = 1) 

#Saves the new patient info into a new file. For safekeeping.
NewPatient2.to_csv('NewPatient2.csv') 



np2 = pd.read_csv('NewPatient2.csv') #New Patient 2 info.

np2.head()#Lets see the new patient's submitter Id to get the DNA Methylation data.

This new patient is not present in the list of patients for group 2 and thus can represent patient 2.

#importing new patient 2 DNA Methylation data.
g2np2 = pd.read_csv('DMG2NP2.txt', delimiter = '\t', encoding = 'utf-8') 

g2np2.isna().sum() #Total sum of all missing values for every column.

len(g2np2) 

The size of the dataset is a lot better than the first patient 2, and can be used to analyze alongside the others in the group. 

g2np2 = g2np2.dropna() #Removing rows of missing values.

g2np2.isna().sum() #Checking that the data is clean and free from NAs.

## Group 2 - Patient 3

g2p3.isna().sum() #Total sum of all missing values for every column.

len(g2p3) #Size of dataset.

g2p3 = g2p3.dropna() #Removing rows of missing values.

g2p3.isna().sum() #Checking that the data is clean and free from NAs.

There seems to be a trend amongst the mising values. Perhaps all datasets have approximately the same value missing from the same column? If so, we can skip a few steps and go straight to removal of missing values.

## Group 2 - Patient 4

g2p4.isna().sum() #Total sum of all missing values for every column.

len(g2p4) #Size of dataset.

g2p4 = g2p4.dropna() #Removing rows of missing values.

g2p4.isna().sum() #Checking that the data is clean and free from NAs.

## Group 2 - Patient 5

g2p5.isna().sum() #Total sum of all missing values for every column.

len(g2p5) #Size of dataset.

g2p5 = g2p5.dropna() #Removing rows of missing values.

g2p5.isna().sum() #Checking that the data is clean and free from NAs.

## Group 2 - Patient 6

g2p6.isna().sum() #Total sum of all missing values for every column.

len(g2p6) #Size of dataset.

g2p6 = g2p6.dropna() #Removing rows of missing values.

g2p6.isna().sum() #Checking that the data is clean and free from NAs.

## Group 2 - Patient 7

g2p7.isna().sum() #Total sum of all missing values for every column.

len(g2p7) #Size of dataset.

g2p7 = g2p7.dropna() #Removing rows of missing values.

g2p7.isna().sum() #Checking that the data is clean and free from NAs.

## Group 2 - Patient 8

g2p8.isna().sum() #Total sum of all missing values for every column.

len(g2p8) #Size of dataset.

g2p8 = g2p8.dropna() #Removing rows of missing values.

g2p8.isna().sum() #Checking that the data is clean and free from NAs.

## Group 2 - Patient 9

g2p9.isna().sum() #Total sum of all missing values for every column.

len(g2p9) #Size of dataset.

g2p9 = g2p9.dropna() #Removing rows of missing values.

g2p9.isna().sum() #Checking that the data is clean and free from NAs.

## Group 2 - Patient 10

g2p10.isna().sum() #Total sum of all missing values for every column.

len(g2p10) #Size of dataset.

g2p10 = g2p10.dropna() #Removing rows of missing values.

g2p10.isna().sum() #Checking that the data is clean and free from NAs.

# Download All Clean Files:

Lets download all the patient's clean datasets.

g2p1.to_csv('g2p1.csv')
g2np2.to_csv('g2p2.csv')
g2p3.to_csv('g2p3.csv')
g2p4.to_csv('g2p4.csv')
g2p5.to_csv('g2p5.csv')
g2p6.to_csv('g2p6.csv')
g2p7.to_csv('g2p7.csv')
g2p8.to_csv('g2p8.csv')
g2p9.to_csv('g2p9.csv')
g2p10.to_csv('g2p10.csv')

Now that all DNA Methylation data for every patient on group 2 is clean, I will use it to perform exploratory data analysis in part 4.
